Implemented the download rendered image feature. Needs a Firefox fix ?
The user can now dowload the previewed (proxy) result in the format of
his choice between PNG, JPG, TIFF or the format of the original input
image if the plugin is a filter.

This is basically done by sending another request to process the image
again with the same renderer settings, and sending the generated file to
the user.

There is still an issued however : the Javascript seems unbale to make
Firefox browsers download the images, it may requires a little
modification.