Can now change the representative image from the plugin gallery.
Added a small icon underneath each example thumbnail to set the default
image for the plugin. Need somme addirinal work to make it more
ergonomic.
Also modified the plugins catalog template accordingly to display the
default image on the plugin, or the first example from the gallery if
there is no default image.
Additionaly, theses images are now loaded as the thumbnails versions.